Using shared test fixtures for test video
