+++ tests/test_stats_manager.py
@@ -72,29 +72,11 @@ from scenedetect.stats_manager import NoMetricsRegistered
 from scenedetect.stats_manager import NoMetricsSet
 
 
-TEST_VIDEO_FILE = 'testvideo.mp4'
-
 # TODO: Replace TEST_STATS_FILES with a @pytest.fixture called generate_stats_file.
 #       It should generate the path to a random stats file for use in a test case.
 TEST_STATS_FILES = ['TEST_STATS_FILE'] * 4
 TEST_STATS_FILES = ['%s_%012d.csv' % (stats_file, random.randint(0, 10**12))
                     for stats_file in TEST_STATS_FILES]
-
-
-@pytest.fixture
-def test_video_file():
-    # type: () -> str
-    """ Fixture for test video file path (ensures file exists).
-
-    Access in test case by adding a test_video_file argument to obtain the path.
-    """
-    if not os.path.exists(TEST_VIDEO_FILE):
-        raise FileNotFoundError(
-            'Test video file (%s) must be present to run test cases' % TEST_VIDEO_FILE)
-    for stats_file in TEST_STATS_FILES:
-        if os.path.exists(stats_file):
-            raise FileExistsError('Existing file would be overwritten by running test, aborting.')
-    return TEST_VIDEO_FILE
 
 
 def test_metrics():
@@ -339,4 +321,3 @@ def test_load_corrupt_stats(test_video_file):
     finally:
         for stats_file in stats_files: stats_file.close()
         for stats_file in TEST_STATS_FILES: os.remove(stats_file)
-
