+++ tests/test_video_manager.py
@@ -44,9 +44,6 @@ directly:  https://github.com/Breakthrough/PySceneDetect/tree/resources/tests
 # pylint: disable=redefined-outer-name
 
 
-# Standard Library Imports
-import os
-
 # Third-Party Library Imports
 import pytest
 import cv2
@@ -64,23 +61,6 @@ from scenedetect.video_manager import VideoDecoderNotStarted
 # TODO: The following exceptions still require test cases:
 from scenedetect.video_manager import VideoFramerateUnavailable
 from scenedetect.video_manager import VideoParameterMismatch
-
-
-# Video file used by test_video_file fixture.
-TEST_VIDEO_FILE = os.path.join(os.path.abspath(os.path.dirname(__file__)), "testvideo.mp4")
-
-
-@pytest.fixture
-def test_video_file():
-    # type: () -> str
-    """ Fixture for test video file path (ensures file exists).
-
-    Access in test case by adding a test_video_file argument to obtain the path.
-    """
-    if not os.path.exists(TEST_VIDEO_FILE):
-        raise FileNotFoundError(
-            'Test video file (%s) must be present to run test cases' % TEST_VIDEO_FILE)
-    return TEST_VIDEO_FILE
 
 
 def test_video_params(test_video_file):
